Made it unlikely (but not impossible) for double-playing a music track
It's a pretty simple change: now, if it selects the same ambient or
combat music as the last music played, it reselects music up to 3
times (as long as that same track is chosen). This makes it extremely
unlikely, but not impossible, for the same track to play twice (for
example, if 3 music choices are available, this reduces the chances
of playing a track twice from 33% to 1%).

I chose this method both because there's nothing wrong per se with
playing a duplicate on occasion, and in case only one music choice is
defined for some reason.