Made Thurion invisible by default, reveal them with FLF Catastrophe event.
This change causes most Thurion assets to not exist on the map at first,
and it causes the Thurion core systems (the only ones not hidden by
default) to be behind a hidden jump. This ensures that the Thurion
cannot be discovered until a mission or event has designated them as
revealed, similar to Sindbad.

This commit also includes adding such a reveal to the FLF Catastrophe
event, which includes a bunch of new dialog explaining to the player
who the Thurion are and pointing out that a Thurion ship is needed to
safely traverse the nebula.