Modified the way info about cargo is displayed to fix an issue.
The previous way showed type of cargo, which is cool in theory, but it
caused later info to be offset and disappear out of the bottom. Since
it's theoretically possible for any number of different kinds of
commodities to exist in one ship, I think it's best not to depend on
being able to list all of them. Better to just say the total amount
of cargo and be done with it.