Replaced the polygon definition of asteroid fields with simple circles.
I kind of get why it was implemented that way: you can define all kinds
of wacky shapes real easily. But in the vast majority of cases don't
need that, and it makes defining asteroid fields needlessly complicated.
In the few cases where complicated shapes might be needed, it's easy
enough to approximate shapes with a series of circles.

The code is much simpler, too.

Display in the map editor has been removed at least temporarily.
Hopefully should add it back in later.