Made the chance of getting hailed for the Baron mission decrease over time.
With this chance, the chance of getting hailed for the Baron mission
starts at 4% the first three times, then after that decreases by half
each time it happens, so it goes to 2%, then 1%, then 0.5%, and so on.
To ensure that the player can always get the mission somehow, the
Toaxis system is a special case where the chance is always 4% no
matter what.