Made "alive" bounties use a custom faction.
This custom faction is enemies with civilians, traders, and Sirii,
and allies with the Pirates, but not enemies with other police
forces. Justification being that since the pirate is wanted alive,
officials who can't disable the target (i.e. non-Sirius) wouldn't
go after it.

As a special exception, the Dvaered and the special faction are
still enemies, since the Dvaered are notably absent-minded on this
sort of thing and also don't offer alive bounties.