Improved start-of-game tutorial, replacing the menu tutorial
I decided to do this because the classic tutorial is quite long, and
yet the intro text I had previously added is inadequate. This new
tutorial is designed to be a nice middle-ground.

As before, the tutorial is optional, offered to you at the start of
the game. Unlike before, you go through an entire mission
demonstrating all the important game mechanics, rather than just
having a bunch of information thrown at you.

This makes the tutorials (which I haven't recommended to anyone in
years) obsolete, so the "Tutorial" button and old tutorial mission
have been removed.