Converted a background's color map to suppress regular warnings.
This is either the background or one of the backgrounds that cause
those regular warnings about a Photoshop color map. I opened the
file in GNU IMP, converted the color map, re-saved it, and then
optimized it with Trimage. The warning is now gone for this
background at least. (I'll do the same for any other background that
I notice causes the warning.)