Fixed friendly/hostile setting.
The previous system was very hodge-podge and had a lot of duplicated
checks, leading to some inconsistent behavior. I've consolidated all
checks to the setHostile / setFriendly / isHostile / isFriendly
functions.

This fixes a bug where a pilot that was set as friendly but became
hostile due to aggression was impossible to target with the "target
nearest hostile" button.