Increased all swivel values to 22 degrees.
The reason for this is to make aiming less difficult and to reduce
the need to use autoaim. I tested with both a Kestrel and a Hyena.
It seems to be a good swivel value.

Set all of them to the same value because I don't think the game was
benefiting from having different swivel values for different
weapons.