Added a "virtual Sindbad" for when Sindbad is not discovered.
This ensures that FLF presence is always in Sigur and surrounding
systems.

I noticed a bug where for some reason, removing this Virtual
Sindbad asset caused the Virtual Empire Unpresence to be removed
too? I'm massively confused by that. "Fixed" by having a redundant
entry under the FLF_base diff but this should be fixed for real
later, whatever the cause may be.