Added an API for allowing missions / events to force certain music.
It works by having a var to force a certain faction, one for ambient
and one for combat. Dangerous and missions *must* have a thing in
the abort() function popping the variables (you don't want it to
bug out and leave you stuck with factional music forever), but it
works!

Potentially useful for something to the effect of defending against
invaders.