Added a quick tutorial event for when you first buy a large ship.
This tutorial event briefly explains how time dilation works when
you buy your first large ship. It's only a single message window,
so it shouldn't be a major bother to those who already know. I've
also integrated it in as a concerned salesperson trying to provide
good customer service to explain the sudden message away.

The message also explains that larger ships should generally use
turrets, since that's related to the Time Dilation issue.

This may help with #1078; the Time Dilation was not explained
particularly well and this ensures that an explanation is always
given when it becomes an important point.