Removed the Commodity Run missions.
These missions don't work anymore because commodities are available
everywhere (just at different prices). There is no way to salvage
this; it will always be possible to take advantage of jacked-up
prices by buying a commodity at the same planet, taking off, and
landing (this is why they don't spawn anymore, the mission itself
prevents this by not allowing the mission in places where the
commodity is available).

This should probably be replaced by some sort of mechanism to make
commodities specially cheap or expensive, just as an event. But for
now, the Commodity Run mission is an unsalvagable deadweight, so
this removes it.