Started adding more systems and assets for the Proteron.
I've added in all the systems that I want to, but not all the assets.
All but one of these systems will have one or two assets added to them
later; I'm just not doing it right now because I'm tired and it's late.

I also modified Protera Sigma to not offer Proteron ships, instead
offloading that task to a Proteron station with a military restriction.