Fixed up lots of style errors, removed travel log at least partly
Please look over this, but I had a go at removing the travel logs.
Didn't do any testing beyond making sure it compiles. I was exhausted
by all the things I was compelled to fix and just had to stop. I
*think* I removed the travel log stuff though.

A note on style: you consistently get a lot of style wrong and it's
a pain to correct. Please take a look at some of the fixes I've made
here and make further adjustments following this.

In addition to the kinds of style errors I fixed, I also noticed you
using printf to print warnings. You should *not* do this; Naev has a
macro specifically designed for warnings, the `WARN` macro. You should
*always* use macros like this (`WARN`, `DEBUG`, or `ERR` depending
on the situation) for a whole host of reasons, not least of which is
that these macros do more than just print a message. You need to fix
all of these uses of `printf`. It's never something you should be using
in Naev code.

🕵️