Added a "slow mode" for accessibility.
This was something we mused about before: the idea of adding a "slow
mode" as an accessibility feature. Slow mode is pretty simple, it
simply halves all speed in space compared to what it would normally
be. This is helpful for people who have difficulty with motor control
or otherwise have trouble running the game at normal speed (much like
an easy mode).