Added forcing FLF combat music on es00.
It's very likely the player isn't hostile to the FLF at this point,
so it would have played pirate music (in some cases) or the default
music. This change causes the mission to force FLF music. It also
makes the mission claim the system it's in, since music_combat_force
needs to be popped when leaving the destination system and we wouldn't
want two missions conflicting over the music_combat_force variable.
It's also probably for the best anyway since it's spawning ships in
the system.