Updated and renamed the "bar template".
It's now called "mission template" and it's much more extensive...
and most importantly, up-to-date. I actually might have gone slightly
overboard here seeing as the template is actually a 100% valid
mission, albeit a simple one. Like, this could legitimately be
added into the game if we wanted.

I think it might be best to delete "complete.lua" and
"computer_template.lua". Two reasons:

1. mission_template is complete, so "complete.lua" is redundant.
2. Beginners shouldn't really be making new computer missions since
   computer missions should usually be generic. Those require a lot
   more care than unique missions.