Added symbols for faction text on the map screen, and for ship stats
This is an accessibility feature for those who are colorblind. You
no longer need color vision to see this information.

Based on the idea we had before to use emojis, but using ASCII instead.
The biggest advantage of this is that it can just be directly inserted
into the text rather than having to worry about positioning the symbols
manually.