Made it so that overlay text is always partially transparent.
This prevents the text from having outlines. A temporary fix for a
problem where the outlines were obscuring other parts of the overlay
map, but ultimately the overlay map should be redesigned so it doesn't
have text clobbering together all over the place. I'm not sure what
the solution is right now though, so fixing it this way for now.