Made the log delete button less destructive.
I didn't notice this before: the log "delete" button deletes logs
(not log entries, but entire logs) by category. That's fine I suppose,
but it was willing to delete ALL logs, which we probably don't want.
I've changed it to always only delete one specific log, with a warning
that all log entries in said log will be deleted and that it cannot
be undone.