Nerfed Operation Black Trinity.
Black Trinity was nearly impossibly hard; the Trinity was very eager
to flee, meaning you had to pummel it with very powerful shots very
quickly, and it had a huge swarm of drones, meaning you had to be
able to deal with them too. The overall result of this was that it
was basically impossible to "win" the mission without a large and
very powerful ship, and without entering from the top (which clearly
isn't the place you're "supposed" to enter). This has been remedied
by substantially nerfing the fight:

* The Trinity runs when its armor is damaged rather than when its
  shield is at 70%.
* The first swarm of drones does not spawn, only the reinforcements
  do.
* The Collective Sml Swarm fleet has been reduced in number slightly
  (3 drones and 1 large drone rather than 4 drones and 2 large). I
  made sure, and this doesn't affect any other missions in any
  noticeable way other than the visuals of ec02 (you'll see four
  drones a swarm instead of six in the cutscene).

Also fixed a typo.