Tweaked the balance of small ships slightly.
I slightly nerfed the Hyena in an attempt to make the Shark a valid
small ship choice, and I slightly nerfed the Pirate Shark as well.
Both now have a speed mod of 5 (instead of 10), and the Hyena now
has an increased armor penalty and reduced thrust mod.

The Hyena and Shark are still mostly identical, but the choice is
less clear now: you have the Hyena, which is faster but weaker and
with less utility / structure slots, or the Shark, which is slower
but stronger and with more slots.