Added a "creditstring" function to numstring.lua.
This is designed to make proper use of ngettext for credits easier.
It does the right thing for translation in combination with using
the numstring function. I've already made use of it in a couple of
places. I'm essentially recommending this for any time Lua code needs
to display a certain number of credits.