Reworked aristodrink text.
The previous text tried to do something that's interesting, but
doesn't work very well for translation. As such, I've replaced this
with a more simple system. In keeping with the "gender choice" theme,
the mission generates three variations: one where the final bartender
is male, one female, and one nonbinary.

This edit also removes some text that was a bit sexist (referring to
a stranger's dress as "sexy", which also casually makes an assumption
that the player finds women attractive).