Conversion of traders to a static faction (like civilians)
This is just a bit of a proposal and a demonstration of what that
entails. I don't really care either way whether or not we go with
this, but just want to make my case that honestly, I don't feel like
traders should be a facion proper with reputation boosts and such.
They should be a static faction like civilians. There's a couple
reasons for this:

1. Traders are all over the galaxy, so the idea of traders having
   their own position in the galaxy seems a bit silly.

2. Traders don't fly military ships and the trader AI isn't designed
   to defend systems from piracy in any way.

3. The player theirself can be a trader and yet not be a part of the
   "space traders guild". Wouldn't the simple act of trading make
   you a space trader?

4. I can't see any direction traders as a "real" faction could go
   that would make sense and be fun.

Anyway, let me know what you think!

🕷️