arnoldsmith2 -> arnoldsmith (commit 3 of 3)
OK, so an explanation: in these three commits, I just swapped the
arnoldsmith image with the unfamiliarman image. Nothing major. I did
it in three commits simply so that the Git history would properly
record it as renaming (and not modifying) the files.

The reason I swapped the images is because part of Arnold Smith's
campaign involves things getting more and more sketchy and the
portrait I originally made for him is dark-skinned, so while this
wasn't my original intention, that could fuel racial prejudice (in
the form of belief that PoC that look honest are actually bad or
something).