Changed all credit shortening suffixes to SI symbols.
B -> G, Q -> P, Qt -> E

This makes the suffixes consistent ("K" is an SI symbol), and also
less English-centric. (SI symbols are an international standard, but
"B", "Q", and "Qt" are representations of the English words "billion",
"quadrillion", and "quintillion". Changing to SI should make the
suffixes more universally understood: G for giga, P for peta, E for
exo. M and T happen to be unchanged, but now stand for Mega and Tera
rather than Million and Trillion.)