Fixed up equip script, improved Proteron kahan balancing.
This replaces the heavier weapons and extra med utility slot with
lighter size and an extra lge structural. Works better that way I
think. The lighter mass (lightest mass of any cruiser) enables it
to be faster than other cruisers while carrying more of a payload.
Given this, it's had its time dilation value decreased as well,
to 175% (compared to the 150% time dilation of corvettes), not
unlike the Kestrel in that sense.