Run the note about incompleteness with every version change.
This also adds the version number to the message.

The reason for using the version *string* is because it's hard to
predict exactly what parts of the version we'll want to differentiate
in the future and updates to the version number should be infrequent
enough to not pose a problem.