Removed remaining sanist language (mostly from source code comments).
Bit of an explanation: I understand that "sanity checks" is a common
term in programming parlance, but this common usage stems from sanism
that has been entrenched into our culture and even into the English
language. By describing bad situations as "insane", you end up
subtly implying that those with mental illness cause problems, even
if you did not intend that connotation. This change removes all those
references and replaces them with terms that both are not sanist and
which are more precise (usually "validity" or "safety").

This also meant that the "in-sanity" tool's name was changed to
"in-valid" (keeping a joke but one that doesn't have sanist origins).
Some function names were also updated.

This also includes updated POT files since I wanted to make sure I
got everything. The only sanist references left now are those in
Autoconf itself (which of course we can't really do anything about),
and possibly words like "stupid", "idiot", "moron", which are also
sanist, but which I haven't specifically looked for yet.