Removal of all unique outfits and the "traveling merchent".
I've made my case on #1224 as to why, but this PR demonstrates exactly
what I want to happen:

I want *all* of this "unique outfit" stuff removed. I don't like it
and don't think it should be a thing. The functionality in the engine
being there is fine, but I don't want us to use it.

Naev isn't an ARPG, and I don't think we should treat it as one.
"Traveling merchants" make sense in RPGs, but in Naev you get the
outfits you want by searching for them and going to the planets that
sell them. I don't want to buy "unique outfits" from a "traveling
merchant". I want to buy entirely ordinary outfits from all over the
galaxy.

You may notice that there's no replacement here, and that's because
I disagree not with the implementation, but with the basic premise.
I don't want unique outfits or "player growth" of any kind beyond just
amassing a bigger and more diverse collection of ordinary ships.