Renamed "dumb"/"dumbfire" rockets to "unguided".
The term "dumbfire" (which doesn't seem to be in common use outside
of a select few video games) comes from the term "dumb", which is
ableist on two counts: firstly, it's sanist because the meaning
"dumbfire" is derived from is "stupid" or "of low intelligence".
Secondly, the term "dumb" actually originates from ableism against
mute people, albeit in the German language rather than English.

The term is imprecise, anyway, so I've changed it to the more precise
and non-ableist term, "unguided".