fix shipstats modification procedure for integers
Before that, i was erroneously pointing into a forbidden region of
memory and writing to it caused a segmetation fault.
This went undetected because of a relative lack of integer stats
tickling this branch of logic.

I discovered the problem when playing with Hidden Jump Detector.
(and now it works)

Also, the problem is obvious if one compares to the same code for
doubles.