Added warnings about more things that could cause rush cargo fails.
Attempting to do a rush cargo mission to a system that's more jumps
away than you can do or that you don't know an optimal route to can
be a bad idea. This has the potential to cause confusion (and we've
had at least one reported case of such confusion), so this adds a
warning for those scenarios.