Made the selection color for jumps and planets white
This is a step forward in colorblind accessibility and starts to
bring some consistency. This wasn't done for pilots because the
brightness contrast between yellow and white is too low. Will need
to think about a proper solution. Maybe show the pilot's name next
to it if selected? That would help.