Rebalanced hulls.
Not much changes here, mostly just a more targeted use of absorb to
create "classes" of absorption, more or less. Weapons will target
heavy classes by having high penetration to counter it. This commit
adjusts the Caesar IV rocket to that standard (it is given 70%
penetration, to counter the 72-80% absorb of superheavy plating).