Added symbols to overlay text, fixed a potential text positioning bug
Random note: it occurs to me that depending on text that's chosen
somewhere else being a particular size is probably not the best thing
to do. Would be better to both decide the text and position it in
the same piece of code so we don't risk discrepancies like the one
I fixed (where one part assumed it was always the system name even
though it could be "Unknown").