Vastly improved the Commodities screen.
This substantially reduces the size of the commodities image array
to give much, much more space for the commodity description. This
allows for upscaling the image there (as in the outfits screen),
improving the economy info header text, and giving so much space for
the commodity description that you could write an essay there.

This also changes the programming of the commodities screen so it
can more easily be adjusted later if need be. Rather than weird
absolute values, it modifies placements based on the image array
size, as with most things.