Adjusted the Ion Cannon (but not the Heavy Ion Cannon yet).
This powers up the Ion Cannon substantially so that it is ever so
slightly stronger and more expensive than a Laser Cannon MK1. This
designates the Ion Cannon for use by ultralight craft. Regular fighters
and Corvettes will have the option to either use this, or use Medusa
Missiles, or use the Heavy Ion Cannon, or some combination.

Untested.