Configured Collective weapons and the first two lances.
Collective weapons intentionally break the rules. The Neutron Disruptor
is only a slight tweak, with double damage at the cost of reduced
range, but the Heavy Neutron Disruptor has been modified to combine
elements of various classes. Literally speaking, it's one of the most
massive weapons in the game now, which is designed to compel the
Heavy Drone to use the higher engine mass limit engine. However, as
compensation, it consumes only slightly more CPU than the Laser Cannon
MK2 while inflicting twice as much damage, placing its actual
damage output more or less in-line with the Ripper Cannon or possibly
even the Heavy Ripper Cannon.

Untested, as usual.