Removed unused stuff in generic.lua, added generic pilot name function.
This is a generic version of the pirate name generator which is based
on a simpler, less silly sounding [noun] [number] rule for use in
pilots that aren't pirates (and thus need to be taken at least
somewhat seriously). Added in preparation for my escorts thing, but
also can be used in bounty and seek and destroy missions.