Fixed auto-boarding, made boarding ships snap to target's speed.
For the auto-boarding fix: it was checking for the player's absolute
speed, which is incorrect. Boarding requires matching the target's
speed, not necessarily coming to a stop (matters in the case of
active boarding targets).

Other than that, caused ships boarding targets to exactly match their
speed vector when doing so, as if they had physically clamped onto
the other ship. Suggested by @UncombedCoconut to avoid lore oddities.