Llama variants...
Added changes for variants as well as a new luxury yacht (which should work with sightseers. These variants are, compared to the stock one; more maneuverable but less tough, tougher but harder to handle, faster and quieter but slower to turn and more frail (in alphabetical order). Plus the Luxury Yacht is much more expensive and obvious with less cargo but much faster and with less inertia.
All have unique engine trails to differentiate them and also in keeping with their theme.