Added logging and shipai cameos to derelict events and missions
As far as I can tell I didn't fluff any of the coding so I hope that the adding of logs and shipai cameos to the derelict events and missions was okay... I also redirected the blackcat and spacefamily mission logs to be ones related to derelicts given that that's how they start. I expect this is not the best solution for neutral events but it makes it clearer for other people to add new ones to the bottom.