Two trader clan home systems
Two trader home systems; Holly and Ivy.
Holly is an extremely ore rich system with a trinary planet system and an asteroid ring outside that, rich in mining prospects. Now the home to a Trader Clan vertically integrated in Ore and Industrial Machinery trading - heavily robotisised
Ivy is a system with a number of terrestrial worlds that were, for a short time, extremely productive in agriculture... until it was discovered, too late, that one of the local flora - locally known as demon kudzu - would overrun everything that the colonists built within a single year. Unfortunately the demon plant had already transplanted itself to every world in the system (at least it is assumed the ice world suffered the same fate) and the system was quarantined until The Incident inspired some (very hungry) traders to find a way to eat the demon kudzu. These days, with very careful quarantine measures, food, water and medicines are traded very successfully - especially given the rock bottom prices the clan charges, especially its endless supply of low quality vegetable replacement packs made from kudzu.