Recreated point-defense turrets
I was guessing that the "resist" entry in the launchers was how destructible the projectile was and have thus added it to all the launchers which describe launching a physical object (i.e. I left the plasma launchers alone) making a number more of the missiles, rockets and grenades destructable, I find it odd they weren't already.
I also created two point defense turrets. I'll explain the reasoning behind their stats:
They have high CPU - tracking objects that small and fast is harder than targetting ships (and traders have few weapons and can spare the CPU cycles so these turrets are less use to combat ships which have armour and shields too).
They are cheap but make up for that in combat effectiveness... 1200 trackmax and 350 range. The amount of heat they produce is disproportionate to their damage energy and speed due to the computation and finely tuned wavelengths specifically aimed at taking out individual types of missiles, rockets and grenades and thus are heat inefficient but then again a trader isn't produce heat from their main guns (like as not) so it's not an issue for a trader (unlike a combat ship).
Their shots penetrate well and damage is higher than standard lasers due to the precision of their shots and specificity of their wavelengths matching their target.

So, these weapons are heat and CPU inefficient but highly capable close in to the ship, firing faster and hitting harder at targets that get close enough to be a danger... indicentally they might well be useful against fighters that do suicide runs or hover over the ship but such sitting ducks were always easy chaff! ; )