Red Star Cargo hulls
So I've reduce the missile jam a little, reduced the cargo space by a third, reduced the cargo mod and inertia benefits from 25 to 10 and made these hulls more expensive. I hope this is a reasonable tradeoff, especially as the CPU costs are just as high and detection is still as high as it was before.
I also added into lore that the Red Star organisation created hem for their humanitarian work since it seems appropriate that an aid agency would do this if necessary given the higher risks their work poses and that (in theory) they shouldn't be being targeted but stray laser shots don't recognise international symbols! ;)
I should note that there are also a couple of changes to tech.xml to allow for future use in the guild stuff but which should probably be there now anyway.