added interactable crewmates
well, they talk to each other, and they can say something to the captain when he presses the talk button. they have internal variables and most of them don't do anything except create memories that they later recall and try to start conversations on the ship. there are two special kinds of crew implemented, an explosives engineer who plants bombs on ships that were boarded and a companion escort, like a guidance counselor for the crew or whatever. The companion escort gains happiness landing on rich planets but otherwise it slowly deteriorates. All crew members' happiness changes based on productive or unproductive conversation and on random events such as space hysteria of varying degrees. The special crew members like the engineer and the companion also gain satisfaction from other things, such as when an engineer's bomb explodes, everyone one the crew with the "violence" interest earns satisfaction, but loses some if a friendly ship is boarded and no bombs are planted. The other crew members are red herring, and there is a red herring with the same "management" skill as the companion escort that doesn't do anything either. All crew members can cost a salary (not implemented yet) or consume from the ships stockpiles, "forcing" the player to buy 1/10th of some commodity (without receiving the commodity) every time it is available. That way you can slightly control crew costs based on location instead of having to tie everything completely to satisfaction or experience points. All crew has a chatter variable that affects how much they talk, I recommend scrutinizing your crew a lot if they start getting irritating, or dialing the chatter to maximum if you want to hear a lot of incoherent ramblings.