Fixes off-by-one error in mission id calculation
Pilot cargo logic assumes that anything with mission id 0 is a non-mission cargo item, but the first mission item that is added becomes a mission item with id 0. This is an off-by-one error where regular commodities and mission commodities can get mixed up, but you will for instance find that if you have only one mission of 10 diamonds to be moved and buy 30 diamonds of commodities, those 30 "normal" diamonds will be locked while you are in space, and you will not be able to jettison them.