Integrate Tracy for profiling
Tracy is an interesting profiler, that is easy to add as a meson
subproject. But requires instrumentation added to functions one wants to
trace.

Not sure if it's worth adding a bunch of calls to this everywhere, but
it's an interesting thing to experiment with. The GPU tracing support in
particular could be worth looking at.