Use `SDL_ConvertSurfaceFormat` to fix rendering of indexed color pngs
The other branch of this `if` statement already did conversion. It makes
sense that it would be needed for indexed color in either case.

I guess conversion should only be done if the existing format cannot be
used? Instead of always calling this.

This issue occurred on Fedora Rawhide, but not other Linux installs. I'm
guessing that would be related to the version of `SDL2_Image` or `libpng`.