Cannibalize new behaviour
 - either gluttony : eat all even if useless.
 - or cannibalize : eat no more than needed, leave at least 1AP.

Gluttony is proposed when either (1) the boarded ship belongs to a _hated_ faction (-30 or less)
or when (2) the player has nothing to gain in cannibalizing.

If player cannibalizes, next click onto "cannibalize" will force gluttony,
as the condition (2) is now met.

In case of gluttony, the players now loses standing for "destroy"ing.
The gluttony immediately closes the board window.

In case of cannibalize, the players now loses standing for "board"ing (looting).

Resized cannibalize button.