Update empire.md
Combined the duplicate paragraphs in the Great Houses section to be one coherent item. Changed the military references to be naval references as that tends to be how all sci-fi views space faring militaries (using naval terms instead of Army or even Air Force terms). Changed some wording to ensure that the Emperor remains gender neutral. Referenced the governmental system as a dictatorship instead of monarchy (monarchies are predicated on one 'royal family' who has power through birthright, dictatorships are the broader term for a single individual being in charge).

Made some spelling, punctuation, and grammar changes. Notably I changed 'the' to lower case when referencing the empire or emperor.